test,minitor: use 64-bit counters

+++ src/test/csrc/monitor.cpp
@@ -9,11 +9,11 @@ enum {
 
 static int g_trapCode = STATE_RUNNING;
 static uint64_t g_trapPC = 0;
-static int g_cycleCnt = 0, g_instrCnt = 0;
+static uint64_t g_cycleCnt = 0, g_instrCnt = 0;
 
 bool is_finish() { return g_trapCode != STATE_RUNNING; }
 
-extern "C" void monitor(int trapCode, uint64_t trapPC, int cycleCnt, int instrCnt) {
+extern "C" void monitor(int trapCode, uint64_t trapPC, uint64_t cycleCnt, uint64_t instrCnt) {
   g_trapCode = trapCode;
   g_trapPC = trapPC;
   g_cycleCnt = cycleCnt;
